+++ app/Http/Controllers/Api/OfferController.php
@@ -260,8 +260,24 @@ class OfferController extends Controller
         return array_merge($data,$page);
     }
 
+    /**
+     * Returns a Sublist of offer-short List (filter on keyword)
+     *
+     * @param String $type
+     * @return \Illuminate\Http\Response
+     */
+    public function getOfferSubListWithKeyword( String $keyword ) {
 
-
+        $keyword= preg_replace('/\s+/', '', $keyword);
+        $offers = Offer::select('offers.*')
+            ->leftJoin('huboffers','offers.id', '=', 'huboffers.offer_id')
+            ->whereRaw("FIND_IN_SET(?, huboffers.keywords) > 0", $keyword)->get();
+        $output = array();
+        foreach ( $offers as $offer ) {
+            $output[] = $this->getReducedOfferJson($offer);
+        }
+        return response()->json($output, 200);
+    }
 
     /**
      * Remodel the output of an offer