+++ Changelog.md
@@ -9,6 +9,10 @@
 - added testroute {api}/offer/paginated/{offerCount}
 - added testroute{api}/list/offer/short/paginated/{offerCount}
 
+## 2021-09-14/ml
+### added
+- added function to get Offers by Keyword
+
 ## 2021-09-03/dk
 ### Changed
 - Upgraded Laravel to 8